// src/app/models/contatos.ts
export class Contatos {
    CodigoContato: number;
    Nome: string;
    Descricao: string;
    Telefone: string;
    Celular: string;
    Email: string;
    EnderecoRua: string;
    EnderecoNumero: string;
    Bairro: string;
    Cidade: string;
    Cep: string;
    CodigoBaseUF: number;
    Municipios: any [];  
    CodigoBaseMunicipio: number;    
    GeoLongitude: number;
    GeoLatitude: number;
}